Fix various QA script lints
Ran Google Fonts QA script, it came up with a few things it didn't like:

* ymax exceeded the usWinAscent value. This was mostly due to extrapolation
of the black weight, and was fixed by reining in the very tallest characters.
Also clipped the darkshade character.

* weightClass of ExtraLight should be 275 to work around GDI failures.

It still complains about the names of the non-regular width fonts, but
there is no guidance about how to handle that in a variable font. I think
traditionally there's just a different family name for each width.